Extract utility class for argument checks in SPI
